Add more eslint changes
[Context]
 - This is part of a stack of commits to improve the itaigi codebase

[This Diff]
 - Renames some js to jsx files. Continues eslint work

[Test Plan]
 - itaigi loads properly on localhost
